# docs/openapi.yaml
openapi: 3.1.0
info:
  title: GSNpeeps API
  version: 0.7.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    Kontrak HTTP GSNpeeps untuk Employee Database, dashboard HR, absensi,
    approval, notifikasi, dan administrasi akses.

    API menggunakan field JSON `snake_case`, Bearer JWT berumur 8 jam, dan
    cross-check sesi aktif di Redis. Tidak ada refresh-token endpoint.
  contact:
    name: GSNpeeps IT Liaison
servers:
  - url: http://{host}:{port}
    description: Development
    variables:
      host:
        default: localhost
      port:
        default: "8080"
  - url: https://{host}
    description: Production
    variables:
      host:
        default: api.janjikupadamu.id
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/lembur:
    post:
      operationId: createOvertimeRequest
      summary: Ajukan lembur
      description: Karyawan mengajukan lembur untuk dirinya sendiri.
      tags: [Overtime]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateOvertimeRequest"
      responses:
        "201":
          description: Pengajuan lembur berhasil dibuat.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdSuccessResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }
    get:
      operationId: listOvertimeRequests
      summary: Daftar pengajuan lembur untuk approval
      description: |
        Atasan melihat pengajuan bawahan langsung; HR melihat seluruh pengajuan
        aktif; Top Management hanya membaca data agregat melalui endpoint rekap.
      tags: [Overtime]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/RequestStatusParam"
        - $ref: "#/components/parameters/DateFromParam"
        - $ref: "#/components/parameters/DateToParam"
      responses:
        "200":
          description: Daftar pengajuan lembur.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OvertimeRequestListResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token berlaku 8 jam. Backend juga memverifikasi session aktif di
        Redis pada setiap request terproteksi.
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
...
    DateFromParam:
      name: tanggal_mulai
      in: query
      schema: { type: string, format: date }
    DateToParam:
      name: tanggal_selesai
      in: query
      schema: { type: string, format: date }
...
    RequestStatusParam:
      name: status
      in: query
      schema: { $ref: "#/components/schemas/RequestStatus" }
...
  responses:
    BadRequest:
      description: Request atau parameter tidak valid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            invalidRequest:
              value:
                success: false
                error: { code: INVALID_REQUEST, message: Request tidak valid }
    Unauthorized:
      description: Kredensial tidak ada, tidak valid, atau kedaluwarsa.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            unauthorized:
              value:
                success: false
                error: { code: UNAUTHORIZED, message: Autentikasi diperlukan }
    Forbidden:
      description: User tidak memiliki kewenangan atau akses row-level.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            forbidden:
              value:
                success: false
                error: { code: FORBIDDEN, message: Akses ditolak }
...
    ValidationError:
      description: Validasi bisnis atau field gagal.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            validation:
              value:
                success: false
                error:
                  code: VALIDATION_ERROR
                  message: Validasi gagal
                  fields: { tanggal_mulai: Tanggal mulai wajib diisi }
...
    TooManyRequests:
      description: Rate limit terlampaui atau akun terkunci sementara.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            rateLimited:
              value:
                success: false
                error: { code: RATE_LIMITED, message: Terlalu banyak request }
    InternalError:
      description: Kesalahan internal.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            internal:
              value:
                success: false
                error: { code: INTERNAL_ERROR, message: Terjadi kesalahan internal }
...
  schemas:
    UUID:
      type: string
      format: uuid
      example: 0193f406-8d2f-7d0c-bf34-61ff192e46de
...
    RequestStatus:
      type: string
      description: Enum machine-readable yang digunakan sama pada database, query, dan response.
      enum: [menunggu_atasan, menunggu_hr, menunggu_top_management, disetujui, ditolak, dibatalkan]
...
    SuccessResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, const: true }
        data: {}
        message: { type: string }
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: VALIDATION_ERROR }
        message: { type: string, example: Validasi gagal }
        fields:
          $ref: "#/components/schemas/ValidationFields"
    ValidationFields:
      type: object
      additionalProperties: { type: string }
      example: { email: Format email tidak valid }
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, const: false }
        error: { $ref: "#/components/schemas/ErrorObject" }
    PaginationMeta:
      type: object
      required: [page, limit, total_data, total_page]
      properties:
        page: { type: integer, minimum: 1 }
        limit: { type: integer, minimum: 1 }
        total_data: { type: integer, minimum: 0 }
        total_page: { type: integer, minimum: 0 }
    IdData:
      type: object
      required: [id]
      properties:
        id: { $ref: "#/components/schemas/UUID" }
    IdSuccessResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/IdData" }
...
    CreateOvertimeRequest:
      type: object
      required: [tanggal, waktu_mulai, waktu_selesai, alasan]
      properties:
        tanggal: { type: string, format: date }
        waktu_mulai: { type: string, format: time }
        waktu_selesai: { type: string, format: time }
        alasan: { type: string, minLength: 10, maxLength: 2000 }
        dokumen_pendukung: { type: string, format: binary, description: Maksimal 5 MB. }
    OvertimeRequestSummary:
      type: object
      required:
        [id, employee_id, nama_karyawan, tanggal, waktu_mulai, waktu_selesai,
         total_jam, status, created_at]
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        employee_id: { $ref: "#/components/schemas/UUID" }
        nama_karyawan: { type: string }
        tanggal: { type: string, format: date }
        waktu_mulai: { type: string, format: time }
        waktu_selesai: { type: string, format: time }
        total_jam: { type: number, minimum: 0 }
        status: { $ref: "#/components/schemas/RequestStatus" }
        created_at: { type: string, format: date-time }
...
    OvertimeRequestListResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          required: [meta]
          properties:
            data: { type: array, items: { $ref: "#/components/schemas/OvertimeRequestSummary" } }
            meta: { $ref: "#/components/schemas/PaginationMeta" }
...